fix(validation): rk4 solver + float64 dims in the task-vs-SPM path
Two independent bugs, both previously invisible because check_matlab_available()
was broken and the whole spm suite skipped. Fixed in order as each unmasked the
next:

1. simulate_task_dcm used the default adaptive dopri5, which underflows
   ("underflow in dt 0.0") on torchdiffeq 0.2.5 / torch 2.x. Commit c0a7616
   already applied solver="rk4" to the Phase 30 recovery sweep for exactly this
   reason; this validation path was missed.

2. export_task_dcm_for_spm wrote DCM.n and DCM.v without dtype=np.float64, so
   scipy.io.savemat stored them as int64 and MATLAB died in
   run_spm_task_dcm.m:55 -- "Integers can only be combined with integers of the
   same class" on spm_Ce(ones(1, DCM.n) * DCM.v). This is the same spm_Ce
   footgun as Phase 32 / a27828b; every sibling exporter in this file already
   carried the cast (and the comment naming it) and only the original task
   exporter lacked it.

With both fixed, MATLAB/SPM12 runs the task DCM to completion. The three
task-vs-SPM tests still fail, but now for a different and clearly-attributed
reason: "Encountered NaN: loss" out of Pyro Trace_ELBO. validation/
run_validation.py is the SVI orchestrator (run_svi at L346/L525), so these are
the SVI divergence problem, not a MATLAB or export defect.

+++ validation/export_to_mat.py
@@ -137,9 +137,13 @@ def export_task_dcm_for_spm(
                 [[f"stim{i + 1}" for i in range(M)]], dtype=object
             ),
         },
-        # Dimensions
-        "n": np.array([[N]]),
-        "v": np.array([[v]]),
+        # Dimensions as float64 (int64 -> spm_Ce footgun, Phase 32 / a27828b).
+        # Without the cast MATLAB fails in run_spm_task_dcm.m:55 with
+        # "Integers can only be combined with integers of the same class"
+        # on spm_Ce(ones(1, DCM.n) * DCM.v). Every sibling exporter already
+        # carried this cast; this one was missed.
+        "n": np.array([[N]], dtype=np.float64),
+        "v": np.array([[v]], dtype=np.float64),
         # Timing
         "TE": np.array([[0.04]]),
         "delays": np.ones((1, N)) * TR / 2,
